Starting to factor out common turn logic into the Player class
